+++ features/step_definitions/miscellaneous_steps.rb
@@ -14,8 +14,13 @@ Given /^my local git branches are: (.*)$/ do |branches|
   ENV['GEORDI_TESTING_GIT_BRANCHES'] = branches.split(", ").join("\n") + "\n"
 end
 
+Given 'there are no stories' do
+  ENV['GEORDI_TESTING_NO_PT_STORIES'] = 'true'
+end
+
 After do
   ENV['GEORDI_TESTING_STAGED_CHANGES'] = 'false'
   ENV['GEORDI_TESTING_GIT_USERNAME'] = nil
   ENV['GEORDI_TESTING_GIT_BRANCHES'] = nil
+  ENV['GEORDI_TESTING_NO_PT_STORIES'] = nil
 end