+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 ### Compatible changes
+* `geordi commit` will continue even if one of the given projects is inaccessible. It will only fail if no stories could be found at all.
 
 ### Breaking changes