+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Breaking changes
 
-## Unreleased
+## 6.0.0 2021-04-06
 
 ### Compatible changes
 * `geordi commit` will continue even if one of the given projects is inaccessible. It will only fail if no stories could be found at all.